test: move tests into same package as tested code
